Prkt seems to be handling a little bit of noise!
Noise is on the order of .001, but for a while I was using it at
.0001 to use the motion model/expected stuff. It currently took
a lot of trash poses, and then converted them to a few pretty good
ones. It's still pretty slow, I'm going to try to rosbag the data
then run the SLAM on the data as after the fact.